rename parent sidebar file
