+++ src/extraction/features/stratigraphy/sidebar/classes/depth_column_entry_sidebar.py
